# qBitrr/openapi.json
{
  "components": {
    "responses": {
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "error": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "description": "Not authenticated"
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "description": "Value of WebUI.Token in the Authorization header. Query ?token= is also accepted by the server but is discouraged.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "OpenAPI description of the qBitrr WebUI HTTP API. When WebUI.AuthDisabled is false, most routes require a Bearer token (WebUI.Token) or a valid browser session after login. Interactive docs: GET /web/docs or GET /api/docs. Human-readable reference: https://feramance.github.io/qBitrr/webui/api/",
    "title": "qBitrr WebUI API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/arr": {
      "get": {
        "operationId": "api_arr_list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Configured Arr instances (/api)",
        "tags": [
          "WebUI"
        ]
      }
    },
...
    "/api/meta": {
      "get": {
        "operationId": "api_meta",
        "parameters": [
          {
            "in": "query",
            "name": "force",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Version info (/api, requires Bearer or session)",
        "tags": [
          "WebUI"
        ]
      }
    },
...
    "/api/status": {
      "get": {
        "operationId": "api_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "qBitrr and Arr status (/api)",
        "tags": [
          "WebUI"
        ]
      }
    },
...
    "/api/torrents/distribution": {
      "get": {
        "operationId": "api_torrents_distribution",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Torrent counts by category and qBit instance",
        "tags": [
          "WebUI"
        ]
      }
    },
...
    "/web/meta": {
      "get": {
        "operationId": "web_meta_public",
        "parameters": [
          {
            "in": "query",
            "name": "force",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [],
        "summary": "Version and auth flags (unauthenticated; WebUI bootstrap)",
        "tags": [
          "System"
        ]
      }
    },
...
    "/web/qbit/categories": {
      "get": {
        "operationId": "web_qbit_categories",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "qBittorrent categories and seeding stats (/web only)",
        "tags": [
          "WebUI"
        ]
      }
    },
    "/web/qbit/overview": {
      "get": {
        "operationId": "web_qbit_overview",
        "parameters": [
          {
            "description": "qBittorrent instance name, or omit/`all` for every instance",
            "in": "query",
            "name": "instance",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Monitored qBittorrent categories with torrent list (/web only)",
        "tags": [
          "WebUI"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}